# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentHR REST API
  description: |
    Welcome to the AgentHR REST API documentation. This API enables you to programmatically automate AI-powered voice communications at scale.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /campaigns:
    get:
      summary: List Campaigns
      description: |
        Retrieve all campaigns with their current status and progress metrics.
        
        <br />
        
        **Campaign Statuses:**
        - `draft` — Created but not started
        - `scheduled` — Waiting for scheduled start time
        - `running` — Actively making calls
        - `paused` — Temporarily stopped
        - `completed` — All contacts processed
        - `cancelled` — Permanently stopped
        
        <br />
        
        **Required Scope:** `campaigns:read`
      operationId: listCampaigns
      tags: [Campaigns]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of campaigns with progress metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create Campaign
      description: |
        Create a new bulk calling campaign with optional contacts CSV upload. The campaign starts in `draft` status and won't execute until you start it.
        
        <br />
        
        **Simplified Workflow:**
        1. Create campaign with contacts CSV → 2. Start campaign
        
        <br />
        
        **Request Format:**
        This endpoint accepts `multipart/form-data` to allow simultaneous campaign creation and contact import.
        
        <br />
        
        **Campaign Data Options:**
        - Send as JSON string in `data` field, OR
        - Send as individual form fields (name, agentId, etc.)
        
        <br />
        
        **CSV Upload (Optional):**
        Include a `contacts` field with your CSV file. Recognized columns:
        - `phone` or `phone_number` — Required
        - `firstName`, `lastName`, `name`, `email`
        - Any other columns become custom fields
        
        <br />
        
        **Configuration Options:**
        - **Calling Windows:** Restrict calls to business hours
        - **Concurrent Calls:** Control call volume (1-100)
        - **Retry Logic:** Automatically retry unanswered calls
        - **Scheduling:** Start immediately or at a future time
        
        <br />
        
        **Required Scope:** `campaigns:write`
      operationId: createCampaign
      tags: [Campaigns]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - agentId
              properties:
                data:
                  type: string
                  description: Campaign data as JSON string (alternative to individual fields)
                  example: '{"name": "Q1 Outreach", "agentId": "550e8400-e29b-41d4-a716-446655440000"}'
                name:
                  type: string
                  description: Campaign name (use either this or data field)
                  example: "Q1 Lead Outreach"
                agentId:
                  type: string
                  format: uuid
                  description: Agent to use for calls
                phoneNumberId:
                  type: string
                  format: uuid
                  description: Phone number to call from
                engine:
                  type: string
                  enum: [elevenlabs, plivo, twilio-openai]
                  description: Telephony engine
                scheduledStartTime:
                  type: string
                  format: date-time
                  description: When to start the campaign
                timezone:
                  type: string
                  example: "America/New_York"
                callWindowStart:
                  type: string
                  pattern: '^\d{2}:\d{2}$'
                  example: "09:00"
                callWindowEnd:
                  type: string
                  pattern: '^\d{2}:\d{2}$'
                  example: "18:00"
                maxConcurrentCalls:
                  type: integer
                  minimum: 1
                  maximum: 100
                  example: 10
                contacts:
                  type: string
                  format: binary
                  description: Optional CSV file with contacts
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
            examples:
              basic:
                summary: Basic campaign (JSON)
                value:
                  name: "Q1 Lead Outreach"
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
              scheduled:
                summary: Scheduled campaign
                value:
                  name: "Appointment Reminders"
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  scheduledStartTime: "2024-01-15T09:00:00Z"
                  timezone: "America/New_York"
                  callWindowStart: "09:00"
                  callWindowEnd: "18:00"
                  maxConcurrentCalls: 10
      responses:
        '201':
          description: Campaign created with optional contact import results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CampaignResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          contacts:
                            type: object
                            description: Present only if CSV was uploaded
                            properties:
                              fileName:
                                type: string
                              totalRows:
                                type: integer
                              contactsAdded:
                                type: integer
                              contactsSkipped:
                                type: integer
                              invalidRows:
                                type: integer
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Q1 Lead Outreach"
                  status: "draft"
                  agentId: "550e8400-e29b-41d4-a716-446655440001"
                  contacts:
                    fileName: "leads.csv"
                    totalRows: 150
                    contactsAdded: 145
                    contactsSkipped: 3
                    invalidRows: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}:
    get:
      summary: Get Campaign Details
      description: |
        Retrieve detailed information about a campaign including:
        - Current status and progress
        - Contact counts (total, called, remaining)
        - Schedule and timing information
        - Configuration settings
        
        <br />
        
        **Required Scope:** `campaigns:read`
      operationId: getCampaign
      tags: [Campaigns]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Campaign details with progress metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /campaigns/{id}/start:
    post:
      summary: Start Campaign
      description: |
        Begin executing a campaign. The system will start placing calls to contacts according to your configuration.
        
        <br />
        
        **Prerequisites:**
        - Campaign must have at least one contact
        - Sufficient credits for estimated calls
        - Associated agent must be active
        
        <br />
        
        **Behavior:**
        - Calls are placed respecting `maxConcurrentCalls` limit
        - Calling windows are honored (no calls outside configured hours)
        - Failed calls are retried based on `retryAttempts` setting
        
        <br />
        
        **Required Scope:** `campaigns:write`
      operationId: startCampaign
      tags: [Campaigns]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Campaign started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /campaigns/{id}/pause:
    post:
      summary: Pause Campaign
      description: |
        Temporarily pause a running campaign. Active calls will complete, but no new calls will be initiated.
        
        <br />
        
        **Resume:** Use the Start Campaign endpoint to resume a paused campaign.
        
        <br />
        
        **Use Cases:**
        - Credit running low
        - Need to add more contacts
        - Business hours ended
        
        <br />
        
        **Required Scope:** `campaigns:write`
      operationId: pauseCampaign
      tags: [Campaigns]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Campaign paused. Active calls will complete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CampaignResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Campaign'
        meta:
          $ref: '#/components/schemas/Meta'
    CampaignListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Campaign'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...